Basically finished
Need to test: expect a lot of errors.

+++ src/finalHPGame/Spell/rangepower/bullet/movable/BulletMovable.java
@@ -7,10 +7,10 @@ public interface BulletMovable {
 	/*Updates the bullets Location by adding speed to the x or y
 	 * ex. bulletsLocation.setX(speed)
 	 * */
-	public void moveBullet(Location bulletLocation, int speed);
+	public void moveBullet(Location bulletLocation, float speed);
 
 	/**Can the bullet still move
 	 * @return true if the bullet can move
 	 * @return false if the bullet has hit it's bounds*/
-	public boolean canMove(Location bulletsLocation, int speed);
+	public boolean canMove(Location bulletsLocation, float speed);
 }
